Set up static site, get login with discord functioning
This one is a bit one, contains a ton of setup for the static site including the use of sb admin 2 for the template and bootstrap. Ditched the idea of using reactor.

Got login with discord functioning well.

Next steps are to get the main dashboard with guild listing working, logout, among the general site pages like commands, setup/installation and landing page.